Refactor QUIC structure, introduce QUIC_Long base class for QUIC Long form packets
